requirements.txt
# 데이터 수집
kafka-python==2.0.2
python-snappy==0.7.3
python-dotenv==1.0.0

# 스파크 관련
pyspark==3.4.1
delta-spark==2.4.0

# 데이터 조작 및 분석
pandas==2.1.0
numpy==1.24.4

# S3, redis
boto3==1.28.17
botocore==1.31.17
redis==5.0.0


# 워크플로우 관리
apache-airflow-providers-apache-spark==4.1.4
apache-airflow-providers-slack==8.0.0
requests==2.31.0
slackclient==2.9.4

